constants: Change pid file to /run
As per BZ#1130939, using yum install-root with RHEL/CentOS 7 lead to an
unusable system.

Yum's pid file is located in /var/run and will create the directory.

When doing an yum install-root of basesystem, the targeted chroot is
empty, /var/run is so a directory that contains yum.pid.

The filesystem package (dependency of basesystem) is then trying to set
a symlink between /var/run and /run. As /var/run is already existing,
the symlink is not setup.

The chroot is then having two separated dirs : /var/run and /run.

At run time, some programs like dbus expected some file to be in
/var/run while some other like systemd expects /run. As a result, dbus &
systemd are no more able to communicate.

This patch just move yum's pid from /var/run to /run. This makes yum not
creating /var/run as a directory and doesn't break the dbus/system
interactions.